# swagger_api_documentation.yaml
openapi: 3.1.0
info:
  title: NirogPath API
  version: 1.0.0
  description: |
    REST API for the **NirogPath Healthcare & OPD Prescription Hub Platform**.
    A comprehensive healthcare management system supporting Patients, Doctors, Clinic Receptionists, and Superuser Admins.
    
    ## Authentication & Security
    Protected endpoints require a HTTP Bearer JWT token passed in the `Authorization` header:
    `Authorization: Bearer <jwt_token>`
    
    Tokens are obtained by authenticating via `POST /api/auth/login` or `POST /api/auth/register`.
    
    ## System Roles & Access Control Matrix
    - **patient**: Browse doctors/hospitals, book consultation slots, pay token deposit via wallet, track live OPD queue position, view e-prescriptions, log daily medication doses, manage family profiles, and ask prescription follow-up questions.
    - **doctor**: Access doctor queue, advance consultation state (`call-next`, `complete`), broadcast running late warnings, set location & availability schedules, write e-prescriptions, reply to patient prescription queries.
    - **receptionist**: Clinic-wide aggregated metrics across doctors, hospital map uploads, walk-in token creation, and patient transfer management.
    - **admin**: Superuser operations — register doctors, create partner hospitals, manage users, and view hospital analytics.
    
    ## User Story Mapping
    | User Story ID | Role | Description | API Endpoints |
    |---|---|---|---|
...
    
...
paths:
...
  /api/bookings/{booking_id}/directions:
    get:
      summary: Get Directions
      description: |
        Return in-hospital directions for a patient's booking.
        Only available when booking status is 'arrived' or 'in_consult'.
      operationId: get_directions_api_bookings__booking_id__directions_get
      parameters:
        - name: booking_id
          in: path
          required: true
          schema:
            type: string
            title: Booking Id
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/reception/verify-qr:
    get:
      summary: Verify Qr
      description: |
        Verify a patient's QR code token. Returns the associated booking.
        Available to any authenticated user (reception, doctor, etc.).
      operationId: verify_qr_api_reception_verify_qr_get
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
            description: QR JWT token to verify
            title: Token
          description: QR JWT token to verify
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/reception/hospital/map:
    post:
      summary: Upload Hospital Map
      description: |
        Upload a hospital floor map image (jpg/png/webp/svg, max 2 MB).
        Saves to backend/uploads/maps/ and updates HOSPITAL_META in-memory.
        Note: map_image_url persists only for the current server session.
      operationId: upload_hospital_map_api_reception_hospital_map_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/Body_upload_hospital_map_api_reception_hospital_map_post"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/uploads/maps/{filename}:
    get:
      summary: Get Map
      description: Serve uploaded hospital map images (public endpoint).
      operationId: get_map_api_uploads_maps__filename__get
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
            title: Filename
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/doctor/location:
    patch:
      summary: Update Doctor Location
      description: |
        Update a doctor's cabin, floor, landmark, and map pin coordinates.
        Also backfills default direction fields if the doctor has none set.
      operationId: update_doctor_location_api_doctor_location_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorLocationIn"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    Body_upload_hospital_map_api_reception_hospital_map_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
        - file
      title: Body_upload_hospital_map_api_reception_hospital_map_post
...
    DoctorLocationIn:
      properties:
        cabin:
          anyOf:
            - type: string
            - type: null
          title: Cabin
        floor:
          anyOf:
            - type: string
            - type: null
          title: Floor
        landmark:
          anyOf:
            - type: string
            - type: null
          title: Landmark
        map_x:
          anyOf:
            - type: number
            - type: null
          title: Map X
        map_y:
          anyOf:
            - type: number
            - type: null
          title: Map Y
      type: object
      title: DoctorLocationIn
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...